Minor clean-up, VehicleIndex bug fixes
